Python: Split expressions base
Move the expressions to `__init__.py` and the visitors to visitors.py

I also played around with a bound.py, but it is so intertwined with the
unbound, that I think it makes more sense to have it into a single file.
